+++ example/lib/typed_examples/streaming.dart
@@ -1,8 +1,9 @@
+import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:janus_client/janus_client.dart';
 import 'package:flutter_webrtc/flutter_webrtc.dart';
-import 'package:janus_client_example/conf.dart';
+import '../conf.dart';
 
 class TypedStreamingV2 extends StatefulWidget {
   @override
@@ -25,6 +26,7 @@ class _StreamingState extends State<TypedStreamingV2> {
   late StateSetter _setState;
   bool isPlaying = true;
   bool isMuted = false;
+  Timer? _bitrateTimer;
 
   showStreamSelectionDialog() async {
     return showDialog(
@@ -65,7 +67,6 @@ class _StreamingState extends State<TypedStreamingV2> {
 
   initJanusClient() async {
     setState(() {
-      rest = RestJanusTransport(url: servermap['janus_rest']);
       ws = WebSocketJanusTransport(url: servermap['janus_ws']);
       client = JanusClient(
         transport: ws,
@@ -119,6 +120,26 @@ class _StreamingState extends State<TypedStreamingV2> {
         setState(() {
           _loader = false;
         });
+
+        // Monitor bitrate every 5 seconds
+        _bitrateTimer = Timer.periodic(Duration(seconds: 5), (timer) async {
+          if (!mounted) {
+            timer.cancel();
+            return;
+          }
+
+          // Get bitrate for the first video stream
+          final bitrate = await plugin.getBitrate();
+          if (bitrate != null) {
+            print('Current bitrate: $bitrate');
+          }
+
+          // Get bitrate for specific mid (if you know the mid)
+          // final specificBitrate = await plugin.getBitrate('v1');
+          // if (specificBitrate != null) {
+          //   print('Bitrate for mid v1: $specificBitrate');
+          // }
+        });
       }
       if (data is StreamingPluginStoppingEvent) {
         destroy();
@@ -136,10 +157,10 @@ class _StreamingState extends State<TypedStreamingV2> {
 
   cleanUpWebRTCStuff() {
     remoteAudioStreams.forEach((key, value) {
-      stopAllTracksAndDispose(value);
+      stopAllTracks(value);
     });
     remoteVideoStreams.forEach((key, value) {
-      stopAllTracksAndDispose(value);
+      stopAllTracks(value);
     });
     remoteAudioRenderers.forEach((key, value) async {
       value.srcObject = null;
@@ -272,6 +293,7 @@ class _StreamingState extends State<TypedStreamingV2> {
   void dispose() async {
     // TODO: implement dispose
     super.dispose();
+    _bitrateTimer?.cancel();
     destroy();
   }
 }